backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
dj-database-url==2.1.0
markdown2==2.4.13
weasyprint==61.2
redis==5.0.1
python-dotenv==1.0.1
openai>=1.30.0
Pillow
gunicorn==21.2.0
whitenoise==6.6.0
